server/requirements/common.txt
django~=3.0.0
djangorestframework~=3.11.0
requests
djangorestframework-jwt
pillow
django-pydenticon
django-cors-headers
